Add copy of ECP presenter as starting point for LUP presenter
