+++ spec/features/admin_view_claim_spec.rb
@@ -1,0 +1,5 @@
+require "rails_helper"
+
+RSpec.feature "Admin view claim" do
+  Journey.all_policies.each { |policy| it_behaves_like "Admin View Claim Feature", policy }
+end
